dead

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: seongjch <seongjch@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/21 08:41:44 by seongjch          #+#    #+#             */
-/*   Updated: 2022/02/10 18:21:44 by seongjuncho      ###   ########.fr       */
+/*   Updated: 2022/02/11 19:39:36 by seongjuncho      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int	count;
+	size_t	count;
 
 	count = 0;
 	while (s[count])
